[TECH] Adds Rubocop

+++ lib/verto/commands/main_command.rb
@@ -1,9 +1,11 @@
+# frozen_string_literal: true
+
 module Verto
   class MainCommand < BaseCommand
-    desc "tag SUBCOMMAND ...ARGS", "manage the repository tags"
+    desc 'tag SUBCOMMAND ...ARGS', 'manage the repository tags'
     subcommand 'tag', Verto::TagCommand
 
-    desc "init", "Initialize a Vertofile in your repository"
+    desc 'init', 'Initialize a Vertofile in your repository'
 
     option :path, type: :string, default: nil
     def init
@@ -14,7 +16,7 @@ module Verto
       Template.render('Vertofile', to: path)
     end
 
-    desc "version", "Shows Verto version"
+    desc 'version', 'Shows Verto version'
 
     def version
       Verto.stdout.puts Verto::VERSION
@@ -23,12 +25,14 @@ module Verto
     private
 
     def validate_current_vertofile!(path)
+      return unless Pathname.new(path).join('Vertofile').exist?
+
       command_error!(
         <<~ERROR
-         Project already have a Vertofile.
-         If you want to generate a new with verto init, delete the current one with: `rm Vertofile`
+          Project already have a Vertofile.
+          If you want to generate a new with verto init, delete the current one with: `rm Vertofile`
         ERROR
-      ) if Pathname.new(path).join('Vertofile').exist?
+      )
     end
   end
 end
